Upgrade to Ant 1.7.1 and exclude version from maven-ant-tasks
This is a personal itch, driven by a desire to use a task
incompatible with 1.6.5 [ant-deb-task].

Testing over the course of the day suggests that the problems
noted here:

http://groups.google.com/group/leiningen/browse_thread/thread/3a0118282d0394b0/a65645fca0f6f73c?lnk=gst&amp;q=ant#a65645fca0f6f

don't seem to be happening anymore.

A brief attempt to use 1.8.1 ended in tears. Without doing much
research, it looks like lancet or leiningen will need to be
updated to work with that version.