Work around a RuntimeMXBean bug.
RuntimeMXBean.getInputArguments is buggy when an input argument contains spaces.
For an input argument of "-Dmyprop=hello world" it returns two arguments, split at the
space: "-Dmyprop=hello" and "world".

This causes eval-in-project to fail in spectacular ways when leiningen.original.pwd
contains spaces or when path to the clojure jar that gets passed to -Xbootclasspath
contains spaces (which it does by default on Windows XP).