Use lein-deb to package leiningen for deployment on deb based systems
Add lein-deb to the dev-dependencies and add the minimal configuration
to build a package that will install correctly on a debian based
system. The deb currently ships all of leiningen's dependencies
as well as leiningen itself, and sticks them into /usr/share/java.
This is actually pretty rude, and will cause problems if
any other package wants to install these dependencies in this location.

We should either a) work to get these dependencies packaged up
as debs and ship them separately or b) install these debs
to a leiningen specific location to avoid conflicts. I'll email
the list to solicit opinions.

I also needed to tweak bin/lein just a little:

1) fix what seems like a bash typo in LEIN_HOME creation
2) add leiningen itself to the classpath