Don't package everything in the deb
Add a depends section to the deb definition and ant jar from the deb, since
these are available via an existing deb package.

I tried to do the same thing for the maven deps via
libmaven2-core-java and libmaven-ant-tasks-java
(http://packages.ubuntu.com/maverick/libmaven2-core-java,
http://packages.ubuntu.com/natty/libmaven-ant-tasks-java)
but ran into issues with the version of maven (2.2.1) available
in the deb repos. Specifically, using 2.2.1 many lein
commands would do some work and then hang. A thread dump
(https://gist.github.com/757708) suggests this has something
to do with edu.emory.mathcs.backport.java.util.concurrent.LinkedBlockingQueue
but my JVM-fu is not sufficient to immediately understand the problem.

Additionally, libmaven-ant-tasks-java is only available in natty, and
would need to be backported or PPA hosted to be available to other
systems.

A better solution would be to have this package depend on an
older version of maven (2.0.10 would be ideal, since that's what
leiningen specifies in project.clj) but I could not find this
in any repos. An even better solution would be to have leiningen
work on 2.2.1, but it's not clear this is worth the effort
at the moment, since lein may move to aether anyway.

With all this in mind I've decided to just package the maven
dependencies with the leiningen deb for now. To avoid conflicts with
other packages, put leiningen's custom dependencies into
/usr/share/java/lein and put all jars in that directory onto
the classpath in bin/lein.