Fix an issue with leiningen's tests on Windows.
Deleting one of the jar files in the sample test project
was failing on Windows, which was causing all sorts of hard to debug
test errors.

Calling System/gc seems to do the trick (at least on Vista),
but it is not really a proper solution.
If anybody has a better idea, please speak up.