Update tests for compile tests
The classpath was changed to use absolute paths for the checkout
dependencies.  They were pointing to the wrong directory since the
working directory would be the leiningen project instead of the
test project.

This adds the absolute paths for test projects back, as just
changing the classpath does not help for the other tasks.