Enforce nREPL dependencies in :repl profile. Fixes #893.
Enforces a nREPL dependency and a clojure-complete dependency for the :repl
profile. Previously, the first found :repl profile from
`project.clj`-`profiles.clj`-default was returned and merged into the real
profile. Now all :repl profiles that are found are merged into the active
profile.

The dependencies are tagged with the ^:displace tag, and as such will be
discarded if a collision occurs.

The changes made `repl-profile-in-project` fail as it did not contain
`leiningen.core.project/empty-dependencies`, which has a :reduce tag in its
metadata for handling dependency merging. The dependency merger expands the
version to its url, hence the change to a url map.

Also removed a TODO from the file, fixed by 27fe39d.