Remove directory part of root-path when copying jar, re #953.
When specifying :filespecs with `:type :path`, copy-to-jar will iterate
over all elements within the path if the path is a directory, and remove
the path from the front of the path when added as a JarEntry. This is an
issue if you specify a file, as the path will then just be an empty
string. This is solved by invoking the newly implemented dir-string and
remove the result of that instead of removing the path itself.