copy-to-jar :bytes needs to unix-path the :path
All of the other copy-to-jar methods call unix-path on (:path spec),
thus ensuring that files added to the jar from a Windows machine end up
with Unix-style path names.  The byte array variant of copy-to-jar is
the one exception though and the result is that tools such as
lein-cljsbuild will leave Windows-style paths in jar files generated
with the "lein jar" command.