Normalize line endings in slurp-resource
slurp-resource previously expected the resource to use Unix-style line
endings, which it would then convert into system-specific line endings
using fix-line-separators.  The problem with that approach is that
templates built on Windows contain resources with CRLF-terminated lines
and so slurp-resource returned CRCRLF-terminated lines.

The new slurp-to-lf function behaves similarly to slurp, but it
normalizes the line endings in the resource to LF.  This ensures that
slurp-resource is able to produce the correct system-specific line
endings.