Set default :jvm-opts in :base profile to :displace
When the default behavior for `:jvm-opts` is additive, it can lead to
unexpected results if a project tries to add its own options which
conflict with the defaults. For example, the following setting:

    :jvm-opts ["-XX:-TieredCompilation"]

Results in a JVM command line like this, in which the project's JVM
options are overridden by the defaults:

    -XX:-TieredCompilation -XX:+TieredCompilation -XX:TieredStopAtLevel=1

By setting :displace metadata on the default :jvm-opts, any project
configuration of :jvm-opts will replace the defaults.