PROFILES.md: Tip to use composite profile
Sometimes we need to override *parts* of a prticular profile on a project-specific base. To a novice
lein user, it is likely not clear how to achieve this. This tip points her in the right direction -
using a composite profile with a "personal" part that can be overriden in `profiles.clj` to change
what is in the "common" part of the profile.