Updated error message for newbies
When you clone a git repo, it gives you the master branch by default.  Not everyone knows to switch to the 'stable' branch, and the error message when you just clone, cd bin, and .lein is misleading.  Tried to hint how to fix this issue because it's really easy if you only know to do it.