Clarified first-time install instructions
Just tried to clarify a few things that confused me as a newbie.  When I actually clicked on the link to use the package manager, it says that sometimes the packaged versions can be old.  I thought that information was helpful, so I added it here.  Also, many projects are best installed by cloning the git repo.  I went down that path next and was left wondering why the 'master' branch wasn't working, so I clarified that you want to download the lein script from the *stable* branch.  Finally, the wording about the lein script bootstrapping itself and taking longer to run was all correct, but I didn't immediately connect the dots that "bootstrapping" meant that the only thing I had to do was to download the script from the right branch and run it.  So I changed "bootstrap" to "installs itself on the first run."