Fix an issue when deploying project to Clojars.
In some setup, people are adding their GitHub usernames in the URL of the repo such that they don't have to provide their usernames when commiting to GitHub, just their password.

This was an issue with Leiningen since it was ignored and the POM file's <scm> section was not properly generated which leaded to a missing GitHub link on their Clojars project page.

The change in the regex of parse-github-url follows two rules of the GitHub username creation:

  (1) alpha-numeric and underscores only
  (2) minimum two characters