An alternative fix for #1721 to clean all profile paths
If we detect a format specifier in the raw :target-path, e.g. a percent
sign, we adjust the :target-path in the project to be cleaned to reflect
the parent directory of the spliced profile.

Thus we avoid activating every single profile for the project,
potentially activating resource-intensive middleware, just to delete a
directory.