Normalize most sequence arguments to vectors
This allows the use of sequence result alias arguments like
:project/source-paths which otherwise serialize to list expressions and
fail on the evaluation side. Attempts to not vec invocation list forms
by ignoring lists which either start with a symbol or recursively start
with an invocation. Not sure if this is a good idea, but it's possible
so I'll try to allow it.

technomancy/leiningen#1835