allow subprocesses to shutdown gracefully addresses 'lein run messes with shutdown hooks #1854'
solution based on this thread in core-libs-dev mailing list:
http://mail.openjdk.java.net/pipermail/core-libs-dev/2014-March/025525.html

Instead of terminating the subprocess outright with Process.destroy, this
patch opens a outputstream to the subprocess stdin and sends Ctrl-C to end
it.  This allows the subprocess to call its own shutdown hooks.

A with-open is NOT used here because after the process terminates, there's
nothing to close and it throws IOException Stream Closed.