Set short timeout on compiler agent thread pool
When namespaces use agents during load time, threads
are started in the agent threadpool which take 60 seconds to
end. This adds up to 60 seconds of stall time after the compile
task runs.

This patch checks to see if clojure.core/set-agent-send-off-executor!
exists, and if so, calls it with a Thread pool that has the keep
alive time set to 100 milliseconds.

Technically this could cause Thread churn, but practically most
namespace loading doesn't fire up enough Threads for thread
instantiation costs to add up to much.