Fix file handle leak in pom.properties FAQ
There was a file handle leak in the code as it was. It would never close the open reader against the jar resource. This will eventually cause JVM to be unable to open new files.

Functional Tests
----------------------

I'm using the code as I wrote it in my own project and there is now no file handle leak.