(#2067) Deprecate 'resolve-dependencies', add managed deps to it
This commit deprecates the `resolve-dependencies` function in
favor of `resolve-managed-dependencies`.  It also conditionally adds
the `:managed-dependencies` key to the execution of the deprecated
function, in hopes of improving backward compatibility for cases
where people are still calling the deprecated version.

It also cleans up a few docstring "TODO"s and adds real docstrings.