Tolerate keywords for version string - fixes #2212
This commit modifies the logic for normalizing the dependency
vectors.  The previous implementation would not tolerate a keyword
for the `version` identifier in the second position in the tuple.
This caused problems with the lein-module plugin, which allows
keywords to be used as the version identifier.

This implementation takes a different approach, and only assumes
that the vector must end with keyword/value option pairs (if any
options are specified).  It extracts those first, and then tolerates
any remaining value to be interpreted as the version identifier
(including keywords).

This commit also adds some more strict checking of the contents
of the dependency vectors; specifically, it will throw an exception
if the options at the end of the vector are not keyword/value pairs,
or if the artifact id in the first position in the dependency tuple
is not a symbol.