Infer default values from .git/config for :scm :git
Previous when :scm :git was used only the values defined in the :scm
map of the project were used so values not defined there were left blank.
This way it was common to end having an scm node in the pom.xml without
a <tag> node, resulting in the dreaded "with this commit" in clojars
pointing to HEAD instead of pointing to the commit that originated the
release.

Now when :scm :git is used, default values for the :scm node are inferred
from .git/config as if "auto" was used, with the difference that any
value defined under :scm in the project takes precedence over those
inferred defaults. This applies to :url, :connection, :developerConnection
and :tag