[Fix #2492] Respect nREPL's configuration files
This change makes leiningen consult nREPL config files, in addition to
its nREPL config mechanisms (env variables and the :repl-options).

The changes are backwards compatible and any configuration coming from
Lein's own mechanisms in going to take precedence over something
specified via nREPL's config files.

In addition to this, we're also making it easy to specify the
transport to use with nREPL. Extremely useful when you have a client
that doesn't support bencode, but works fine with some of the other
available transports.