Display dynamically the uri-scheme for the nREPL transport
The nREPL dep was bumped to 0.5.2, as it introduced the
uri-scheme function.

I pondered where exactly to put the fallback to bencode and in the end
I've opted to put it here. In hindsight maybe it should have been in
uri-scheme (passing it nil could have returned nrepl as the uri
scheme), but that's not a big deal.