DEPLOY.md - add GPG Troubleshooting
Added troubleshooting tips for two problems I run into and had troubles fixing. The cause of the former is mentioned elsewhere in the docs but when you run into it, a Troubleshooting section is the natural place to look for solution. The latter is caused by inconsistent/unclear docs that mention `lein deploy clojars` but either do not have clojars explicitly in the repositories (because it is baked in, I assume) or have it under "releases". Perhaps some clarification could be made there as well. However we should certainly demonstrate how to make gpg auth work with `lein deploy clojars` since that is what e.g. the tutorial uses.