Remove broken excluded profiles normalization
Commit 328b2182b08e1f6e3d72d6bf7ccf2847de1b1224 added two lines of code
apparently meant to do normalization of excluded profiles. However, the
purpose of this is unexplained and indeed that code is broken.

*Keywords* in exclude-profiles (the normal case) always pass through
"normalize" unchanged; profile maps in exclude-profiles get accidentally
mapcatted to individual map entries, which breaks the excluded profiles
result.

The ineffective/broken code path was almost never taken, and so is
removed here.