Omit unnecessary empty default keys in profile map setup
The setup of profile map defaults always results in the inclusion of the
keys [:jvm-opts :eval-in :dependencies], also when they are not
specified. There is no reason for exactly these keys to be present in
all profile maps, they can be omitted when not given (from the history
it would seem that the possibility of omitting them was simply
overlooked).