Memoize application of middleware
This change introduces memoization of middleware which is meant to
improve performance when using middleware that is expensive or slow to
apply. The performance improvement is particular significant when
running tasks like "install", "jar" or "pom" which involve repeated
merging and unmerging of profiles, as both of these operations involve
reapplying middleware.

To avoid issues with embedded use in long running applications, this
memoization is only enabled by default when running `lein` on the
command line. Memoization can also be disabled by specifying
`:memoize-middleware? false` in the project map or user profile.