Upgrade Apache httpclient
This upgraded version includes a patch for CVE-2020-13956. The version
of httpcore has been updated to match the version preferred by
httpclient itself (see [1][]).

Regenerating `bootclasspath-deps.clj` according to the instructions in
the file also picked up a group ID change for `sjacket` which appears to
correspond to the change made in #2701.

[1]: https://github.com/apache/httpcomponents-client/blob/rel/v4.5.13/pom.xml#L69