Update new task to look for templates the new way.
Previously we had hard-coded the artifact-id to "lein-template" and
used the template name as the group-id, but this is no longer allowed
according to the Clojars verified group policy.

This style is still supported for older templates, but if a template
name contains a slash then the new style is used.