Catch pathological pedantic exception after it happens.
Instead of trying to detect it ahead of time, emit a warning after we
detect the out of memory exception.

This means that it could take a while to notice that there's a
problem, but at least we don't trigger an issue when there wouldn't be
a problem like we were doing previously in 2.9.7.