Set pom-scope during init-profile.
Previously we set pom-scope early on before the profiles had been
expanded. This caused metadata to be lost when a profile like `:dev
[:whatever]` was expanded; the dependencies from :whatever would not
have the pom-scope metadata applied to them.

This caused dev dependencies to be leaked into poms as regular
dependencies. We bypassed that by excluding composite profiles
entirely, but that was also wrong. This fix causes them to be
preserved with the correct metadata attached so that they end up as
being test-scoped in the poms emitted.