Remove excluded profiles after the include-profiles have been expanded
We want to expand the profiles earlier on, but if we do this we can
lose the metadata required for pom-scoping. So instead we do it after
saving off the include-profiles-meta but before actually looping over.

I don't think this actually caused bugs because we were dissoc'ing the
profiles out of the profile map, but it would cause a spurious warning
that the profile in question was not found.