Fix the fallback support for the old release plugin.
When we introduced the built-in release task, it shadowed an existing
"release" plugin, so we added support to prefer the plugin if it was
found.

We had a check in here to avoid an infinite loop of the actual release
task loading the actual release task forever, but it was checking
based on the filename, not the file contents.